# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Hoteles",
        "description": "API para gestionar hoteles, habitaciones, huéspedes y servicios",
        "version": "1.0.0"
    },
    "paths": {
        "/api/guest": {
            "get": {
                "tags": [
                    "Huéspedes"
                ],
                "summary": "Obtener lista de huéspedes",
                "description": "Devuelve una lista de huéspedes con la posibilidad de filtrar por nombre, apellido, pasaporte o DNI, fecha de entrada, fecha de salida e id de la habitación, y paginar los resultados.",
                "operationId": "bdd3e88cba3e6bcf18a3beab32adff44",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtrar por nombre del huésped",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "surname",
                        "in": "query",
                        "description": "Filtrar por apellido del huésped",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "passportID",
                        "in": "query",
                        "description": "Filtrar por DNI o pasaporte del huésped",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "checkinDate",
                        "in": "query",
                        "description": "Filtrar por fecha de entrada del huésped",
                        "required": false,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "checkoutDate",
                        "in": "query",
                        "description": "Filtrar por fecha de salida del huésped",
                        "required": false,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "roomId",
                        "in": "query",
                        "description": "Filtrar por id de la habitación",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de huéspedes obtenida"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    }
                }
            },
            "post": {
                "tags": [
                    "Huéspedes"
                ],
                "summary": "Crear un nuevo huésped",
                "description": "Crea un nuevo huésped con los datos introducidos",
                "operationId": "174c3e57f2db9deb6567d79efd972a70",
                "requestBody": {
                    "description": "Datos necesarios para crear un huésped",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "surname": {
                                        "type": "string"
                                    },
                                    "passportID": {
                                        "type": "string"
                                    },
                                    "checkinDate": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "checkoutDate": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "roomId": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Huésped creado"
                    },
                    "400": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/guest/{id}": {
            "get": {
                "tags": [
                    "Huéspedes"
                ],
                "summary": "Obtener un huésped",
                "operationId": "d85743487481a82c98b04b8b0f09df98",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del huésped",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Huésped encontrado"
                    },
                    "404": {
                        "description": "Huésped no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Huéspedes"
                ],
                "summary": "Actualizar un huésped",
                "description": "Actualiza un huésped con los datos introducidos",
                "operationId": "d10c7f583afe5a0e1cdb1b511a59faf1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del huésped a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos que se pueden actualizar del huésped",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "surname": {
                                        "type": "string"
                                    },
                                    "passportID": {
                                        "type": "string"
                                    },
                                    "checkinDate": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "checkoutDate": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "roomId": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Huésped actualizado"
                    },
                    "404": {
                        "description": "Huésped no encontrado"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Huéspedes"
                ],
                "summary": "Eliminar un huésped",
                "description": "Elimina un huésped por su ID.",
                "operationId": "883beca2e205c4f94d5d8768c6af3edb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del huésped",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Huésped eliminado"
                    },
                    "404": {
                        "description": "Huésped no encontrado"
                    }
                }
            }
        },
...
    },
...
}